# doc/api.yml
openapi: 3.0.3
info:
  title: HEI Graduate Management API
  description: >
    API for managing students, teachers, courses, grades, transcripts and
    graduate lists across the 3-year EL/TN tracks at HEI.
  version: "latest"
servers:
  - url: /
...
paths:
...
  /cohorts:
    get:
      tags: [cohorts]
      operationId: listCohorts
      description: Reference data - any authenticated user can list cohorts.
      responses:
        '200':
          description: List of cohorts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CohortResponse' }
    post:
      tags: [cohorts]
      operationId: createCohort
      x-required-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CohortCreateRequest' }
      responses:
        '201':
          description: Cohort created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CohortResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /cohorts/{cohortId}:
    parameters:
      - $ref: '#/components/parameters/CohortId'
    get:
      tags: [cohorts]
      operationId: getCohort
      description: Reference data - any authenticated user can view a cohort.
      responses:
        '200':
          description: Cohort details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CohortResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [cohorts]
      operationId: updateCohort
      description: >
        Partially updates a cohort (label, entry years). Only the fields
        present in the body are applied; absent fields keep their value.
      x-required-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CohortUpdateRequest' }
      responses:
        '200':
          description: Cohort updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CohortResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
...
    CohortId:
      name: cohortId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
      required: [code, message]
...
    CohortResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        entryYear: { type: integer }
        expectedGraduationYear: { type: integer }
      required: [id, label, entryYear, expectedGraduationYear]
    CohortCreateRequest:
      type: object
      properties:
        label: { type: string }
        entryYear: { type: integer }
        expectedGraduationYear: { type: integer }
      required: [label, entryYear, expectedGraduationYear]
    CohortUpdateRequest:
      type: object
      description: >
        Partial update - only the fields present in the body are applied,
        so no property is required.
      properties:
        label: { type: string }
        entryYear: { type: integer }
        expectedGraduationYear: { type: integer }
...